Add SimCardBerTlvFile, SimCardNoInfoFile classes, EF_NO_INFO and EF_BER_TLV

+++ SIMLibrary/src/de/srlabs/simlib/FileManagement.java
@@ -48,7 +48,7 @@ public class FileManagement {
                 }
             }
 
-            switch ((byte) selectResponse.getFileType()) { // fileType byte
+            switch (selectResponse.getFileType()) { // fileType byte
                 case SimCardFile.MF:
                     selectedFile = new SimCardMasterFile(selectResponse);
                     if (DEBUG) {
